InCallUI: move 'more' menu button to top-right
This menu includes call recording and add to blacklist functionality.
Don't clutter up the button area and move this elsewhere.
