+++ res/layout/call_card_content.xml
@@ -58,6 +58,23 @@
 
     </LinearLayout>
 
+    <!--  More menu overflow button for call recording, blacklisting
+          only visible during the call. It sits above the primary call info
+          container in the top-right hand corner. -->
+    <ImageButton
+            android:id="@+id/moreMenuButton"
+            style="@style/InCallButton"
+            android:paddingTop="16dp"
+            android:paddingRight="16dp"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:background="@drawable/overflow_thumbnail"
+            android:layout_alignParentRight="true"
+            android:layout_alignParentTop="true"
+            android:elevation="@dimen/primary_call_overflow_elevation"
+            android:rotation="90"
+            android:contentDescription="@string/onscreenMoreMenuText"/>
+
     <!-- Contact photo for primary call info -->
     <ImageView android:id="@+id/photo"
         android:layout_below="@id/primary_call_info_container"
