+++ src/com/android/incallui/CallButtonFragment.java
@@ -39,7 +39,6 @@ import android.widget.PopupMenu;
 import android.widget.Toast;
 import android.widget.PopupMenu.OnDismissListener;
 import android.widget.PopupMenu.OnMenuItemClickListener;
-import com.android.internal.telephony.util.BlacklistUtils;
 
 import java.util.ArrayList;
 
@@ -66,7 +65,6 @@ public class CallButtonFragment
     private ImageButton mOverflowButton;
     private ImageButton mAddParticipantButton;
     private ImageButton mManageVideoCallConferenceButton;
-    private ImageButton mMoreMenuButton;
 
     private PopupMenu mAudioModePopup;
     private boolean mAudioModePopupVisible;
@@ -131,24 +129,6 @@ public class CallButtonFragment
         mManageVideoCallConferenceButton = (ImageButton) parent.findViewById(
             R.id.manageVideoCallConferenceButton);
         mManageVideoCallConferenceButton.setOnClickListener(this);
-
-        mMoreMenuButton = (ImageButton) parent.findViewById(R.id.moreMenuButton);
-        if (mMoreMenuButton != null) {
-            boolean canRecordCalls = ((InCallActivity)getActivity()).isCallRecorderEnabled();
-            boolean blacklistEnabled = BlacklistUtils.isBlacklistEnabled(getActivity());
-            if (canRecordCalls || blacklistEnabled) {
-                mMoreMenuButton.setOnClickListener(this);
-                final ContextThemeWrapper contextWrapper = new ContextThemeWrapper(getActivity(),
-                        R.style.InCallPopupMenuStyle);
-                mMoreMenu = new MorePopupMenu(contextWrapper, mMoreMenuButton /* anchorView */);
-                mMoreMenu.getMenuInflater().inflate(R.menu.incall_more_menu, mMoreMenu.getMenu());
-                mMoreMenu.setOnMenuItemClickListener(this);
-
-                mMoreMenuButton.setOnTouchListener(mMoreMenu.getDragToOpenListener());
-            } else {
-                mMoreMenuButton.setVisibility(View.GONE);
-            }
-        }
 
         return parent;
     }
@@ -228,9 +208,6 @@ public class CallButtonFragment
             case R.id.manageVideoCallConferenceButton:
                 onManageVideoCallConferenceClicked();
                 break;
-            case R.id.moreMenuButton:
-                mMoreMenu.show();
-                break;
             default:
                 Log.wtf(this, "onClick: unexpected");
                 break;
@@ -259,7 +236,6 @@ public class CallButtonFragment
         mOverflowButton.setEnabled(isEnabled);
         mAddParticipantButton.setEnabled(isEnabled);
         mManageVideoCallConferenceButton.setEnabled(isEnabled);
-        mMoreMenuButton.setEnabled(isEnabled);
     }
 
     @Override
@@ -575,19 +551,7 @@ public class CallButtonFragment
                 mode = AudioState.ROUTE_BLUETOOTH;
                 break;
 
-            case R.id.menu_start_record:
-                ((InCallActivity)getActivity()).startInCallRecorder();
 
-                return true;
-
-            case R.id.menu_stop_record:
-                ((InCallActivity)getActivity()).stopInCallRecorder();
-
-                return true;
-
-            case R.id.menu_add_to_blacklist:
-                getPresenter().blacklistClicked(getActivity());
-                return true;
 
             default:
                 Log.e(this, "onMenuItemClick:  unexpected View ID " + item.getItemId()
@@ -847,32 +811,5 @@ public class CallButtonFragment
                 getContext().getContentResolver(),
                 android.provider.Settings.Secure.PREFERRED_TTY_MODE,
                 TelecomManager.TTY_MODE_OFF) != TelecomManager.TTY_MODE_OFF);
-    }
-
-    private class MorePopupMenu extends PopupMenu {
-        public MorePopupMenu(Context context, View anchor) {
-            super(context, anchor);
-        }
-
-        @Override
-        public void show() {
-            final Menu menu = getMenu();
-            final MenuItem startRecord = menu.findItem(R.id.menu_start_record);
-            final MenuItem stopRecord = menu.findItem(R.id.menu_stop_record);
-
-            boolean isRecording = ((InCallActivity)getActivity()).isCallRecording();
-            boolean isRecordEnabled = ((InCallActivity)getActivity()).isCallRecorderEnabled();
-
-            boolean startEnabled = !isRecording && isRecordEnabled;
-            boolean stopEnabled = isRecording && isRecordEnabled;
-
-            startRecord.setVisible(startEnabled);
-            startRecord.setEnabled(startEnabled);
-
-            stopRecord.setVisible(stopEnabled);
-            stopRecord.setEnabled(stopEnabled);
-
-            super.show();
-        }
     }
 }
