+++ src/com/android/incallui/CallButtonPresenter.java
@@ -257,29 +257,6 @@ public class CallButtonPresenter extends Presenter<CallButtonPresenter.CallButto
         videoCall.sendSessionModifyRequest(videoProfile);
     }
 
-    public void blacklistClicked(final Context context) {
-        if (mCall == null) {
-            return;
-        }
-
-        final String number = mCall.getNumber();
-        final String message = context.getString(R.string.blacklist_dialog_message, number);
-
-        new AlertDialog.Builder(context)
-            .setTitle(R.string.blacklist_dialog_title)
-            .setMessage(message)
-            .setPositiveButton(R.string.pause_prompt_yes, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    Log.d(this, "hanging up due to blacklist: " + mCall.getId());
-                    TelecomAdapter.getInstance().disconnectCall(mCall.getId());
-                    BlacklistUtils.addOrUpdate(context, mCall.getNumber(), BlacklistUtils.BLOCK_CALLS, BlacklistUtils.BLOCK_CALLS);
-                }
-            })
-            .setNegativeButton(R.string.pause_prompt_no, null)
-            .show();
-    }
-
     public void showDialpadClicked(boolean checked) {
         Log.v(this, "Show dialpad " + String.valueOf(checked));
         getUi().displayDialpad(checked /* show */, true /* animate */);
