+++ src/com/android/incallui/CallCardFragment.java
@@ -41,9 +41,12 @@ import android.telecom.VideoProfile;
 import android.telephony.PhoneNumberUtils;
 import android.text.format.DateUtils;
 import android.text.TextUtils;
+import android.view.ContextThemeWrapper;
 import android.view.Display;
 import android.view.Gravity;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.View.OnLayoutChangeListener;
@@ -57,12 +60,14 @@ import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+import android.widget.PopupMenu;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import android.telecom.AudioState;
 
 import com.android.contacts.common.widget.FloatingActionButtonController;
+import com.android.internal.telephony.util.BlacklistUtils;
 import com.android.phone.common.animation.AnimUtils;
 
 import java.util.List;
@@ -93,6 +98,8 @@ public class CallCardFragment extends BaseFragment<CallCardPresenter, CallCardPr
     private View mCallNumberAndLabel;
     private ImageView mPhoto;
     private TextView mElapsedTime;
+    private ImageButton mMoreMenuButton;
+    private MorePopupMenu mMoreMenu;
 
     // Container view that houses the entire primary call card, including the call buttons
     private View mPrimaryCallCardContainer;
@@ -233,6 +240,20 @@ public class CallCardFragment extends BaseFragment<CallCardPresenter, CallCardPr
         mCallButtonsContainer = view.findViewById(R.id.callButtonFragment);
         mInCallMessageLabel = (TextView) view.findViewById(R.id.connectionServiceMessage);
         mProgressSpinner = view.findViewById(R.id.progressSpinner);
+
+        mMoreMenuButton = (ImageButton) view.findViewById(R.id.moreMenuButton);
+        final ContextThemeWrapper contextWrapper = new ContextThemeWrapper(getActivity(),
+                R.style.InCallPopupMenuStyle);
+        mMoreMenu = new MorePopupMenu(contextWrapper, mMoreMenuButton /* anchorView */);
+        mMoreMenu.getMenuInflater().inflate(R.menu.incall_more_menu, mMoreMenu.getMenu());
+        mMoreMenuButton.setOnTouchListener(mMoreMenu.getDragToOpenListener());
+        mMoreMenuButton.setOnClickListener(new OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mMoreMenu.show();
+            }
+        });
+
 
         mFloatingActionButtonContainer = view.findViewById(
                 R.id.floating_end_call_action_button_container);
@@ -540,6 +561,7 @@ public class CallCardFragment extends BaseFragment<CallCardPresenter, CallCardPr
                 isGatewayCall, isWaitingForRemoteSide);
 
         updateVBbyCall(state);
+        updateMoreMenuByCall(state);
 
         Log.v(this, "setCallState " + callStateLabel);
         Log.v(this, "DisconnectCause " + disconnectCause.toString());
@@ -1164,6 +1186,41 @@ public class CallCardFragment extends BaseFragment<CallCardPresenter, CallCardPr
         mVBNotify.show();
     }
 
+    private void updateMoreMenuByCall(int state) {
+        if (mMoreMenuButton == null) {
+            return;
+        }
+
+        final Menu menu = mMoreMenu.getMenu();
+        final MenuItem startRecord = menu.findItem(R.id.menu_start_record);
+        final MenuItem stopRecord = menu.findItem(R.id.menu_stop_record);
+        final MenuItem addToBlacklist = menu.findItem(R.id.menu_add_to_blacklist);
+
+        boolean isRecording = ((InCallActivity)getActivity()).isCallRecording();
+        boolean isRecordEnabled = ((InCallActivity)getActivity()).isCallRecorderEnabled();
+
+        boolean startEnabled = !isRecording && isRecordEnabled && state == Call.State.ACTIVE;
+        boolean stopEnabled = isRecording && isRecordEnabled && state == Call.State.ACTIVE;
+
+        boolean blacklistVisible = BlacklistUtils.isBlacklistEnabled(getActivity())
+                && Call.State.isConnectingOrConnected(state);
+
+        startRecord.setVisible(startEnabled);
+        startRecord.setEnabled(startEnabled);
+
+        stopRecord.setVisible(stopEnabled);
+        stopRecord.setEnabled(stopEnabled);
+
+        addToBlacklist.setVisible(blacklistVisible);
+        addToBlacklist.setEnabled(blacklistVisible);
+
+        if (mMoreMenu.getMenu().hasVisibleItems()) {
+            mMoreMenuButton.setVisibility(View.VISIBLE);
+        } else {
+            mMoreMenuButton.setVisibility(View.GONE);
+        }
+    }
+
     private void updateVBbyCall(int state) {
         updateVBButton();
 
@@ -1253,4 +1310,31 @@ public class CallCardFragment extends BaseFragment<CallCardPresenter, CallCardPr
             }
         }
     };
+
+    private class MorePopupMenu extends PopupMenu implements PopupMenu.OnMenuItemClickListener {
+        public MorePopupMenu(Context context, View anchor) {
+            super(context, anchor);
+            setOnMenuItemClickListener(this);
+        }
+
+        @Override
+        public boolean onMenuItemClick(MenuItem item) {
+            switch(item.getItemId()) {
+                case R.id.menu_start_record:
+                    ((InCallActivity)getActivity()).startInCallRecorder();
+
+                    return true;
+
+                case R.id.menu_stop_record:
+                    ((InCallActivity)getActivity()).stopInCallRecorder();
+
+                    return true;
+
+                case R.id.menu_add_to_blacklist:
+                    getPresenter().blacklistClicked(getActivity());
+                    return true;
+            }
+            return true;
+        }
+    }
 }
